+++ package.json
@@ -11,6 +11,9 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "@types/babel__core": "^7.20.5",
+    "@types/babel__generator": "^7.6.8",
+    "@types/babel__traverse": "^7.20.6",
     "body-parser": "^1.20.2",
     "express": "^4.19.2",
     "node": "^18.20.4"
