+++ src/reservations/model.ts
@@ -1,27 +1,4 @@
-/* Mock reservations table. Built to simulate a database in local memory. */
-const reservations = [
-  {
-    reservationid: 1,
-    username: "johndoe",
-    propertyid: 1,
-    startdate: "2020-01-01",
-    enddate: "2020-01-07",
-  },
-  {
-    reservationid: 2,
-    username: "brucewayne",
-    propertyid: 2,
-    startdate: "2024-01-01",
-    enddate: "2024-01-07",
-  },
-  {
-    reservationid: 3,
-    username: "sentinel",
-    propertyid: 3,
-    startdate: "2020-01-01",
-    enddate: "2020-01-07",
-  },
-];
+import reservations from "./reservations";
 
 /* Mock database model. Built to be replaced with a mongoose model, as explained below. */
 class ReservationsModel {
