+++ src/reservations/reservations.ts
@@ -1,0 +1,26 @@
+/* Mock reservations table. Built to simulate a database in local memory. */
+const reservations = [
+  {
+    reservationid: 1,
+    username: "johndoe",
+    propertyid: 1,
+    startdate: "2020-01-01",
+    enddate: "2020-01-07",
+  },
+  {
+    reservationid: 2,
+    username: "brucewayne",
+    propertyid: 2,
+    startdate: "2024-01-01",
+    enddate: "2024-01-07",
+  },
+  {
+    reservationid: 3,
+    username: "sentinel",
+    propertyid: 3,
+    startdate: "2020-01-01",
+    enddate: "2020-01-07",
+  },
+];
+
+export default reservations;
