+++ src/users/model.ts
@@ -1,24 +1,4 @@
-/* Mock user table. Built to be replaced with a mongoose DB */
-const users = {
-  johndoe: {
-    username: "johndoe",
-    name: "John Doe",
-    email: "johndoe@gmail.com",
-    role: "USER",
-  },
-  brucewayne: {
-    username: "brucewayne",
-    name: "Bruce Wayne",
-    email: "notbatman@arkham.com",
-    role: "USER",
-  },
-  sentinel: {
-    username: "sentinel",
-    name: "Sentinel",
-    email: "sentinel@admin.org",
-    role: "ADMIN",
-  },
-};
+import users from "./users";
 
 /* Mock database model. Built to be replaced with a mongoose model as described below. */
 class UsersModel {
