+++ tsconfig.json
@@ -7,5 +7,5 @@
       "sourceMap": true,
       "outDir": "dist"
     },
-    "lib": ["es2015"]
-  }
+    "lib": ["es2015"],
+}
